Yes this commit is a mess. Yes I did start updating to 1.21. Yes everything is broken.

+++ src/main/java/io/github/togar2/pvp/enchantment/enchantments/ImpalingEnchantment.java
@@ -4,7 +4,7 @@ import io.github.togar2.pvp.enchantment.CustomEnchantment;
 import io.github.togar2.pvp.entity.EntityGroup;
 import io.github.togar2.pvp.utils.CombatVersion;
 import net.minestom.server.entity.EquipmentSlot;
-import net.minestom.server.item.Enchantment;
+import net.minestom.server.item.enchant.Enchantment;
 
 public class ImpalingEnchantment extends CustomEnchantment {
 	public ImpalingEnchantment(EquipmentSlot... slotTypes) {
@@ -12,7 +12,7 @@ public class ImpalingEnchantment extends CustomEnchantment {
 	}
 	
 	@Override
-	public float getAttackDamage(short level, EntityGroup group, CombatVersion version) {
+	public float getAttackDamage(int level, EntityGroup group, CombatVersion version) {
 		return group == EntityGroup.AQUATIC ? (float) level * 2.5F : 0.0F;
 	}
 }
